upgardes do sth
upgardes machen was und mehr icon sprites

+++ My project/Assets/Enemy.cs
@@ -66,7 +66,7 @@ public class Enemy : MonoBehaviour
         }
     }
 
-    public void takeDam(int dam)
+    public void takeDam(float dam)
     {
         health -=  dam;
     }
